backend/requirements.txt
fastapi
uvicorn
paddlepaddle
paddleocr
opencv-python
pillow
numpy
scikit-learn
